bot/requirements.txt
pyautogui
opencv-python
Pillow
